\name{den}
\alias{den}
\docType{data}
\title{ICA ocular corrected obeject.}
\description{ICA ocular corrected obeject.}
\usage{data(den)}
\format{
  The format is:
List of 7
 $ eeg      :'data.frame':	1920 obs. of  43 variables:
  ..$ Fp1            : num [1:1920] 1.19 1.29 -0.7 -2.13 -1.82 ...
  ..$ AF3            : num [1:1920] -0.305 -0.329 0.179 0.545 0.465 ...
  ..$ F7             : num [1:1920] -0.263 -0.284 0.154 0.47 0.401 ...
  ..$ F3             : num [1:1920] -0.328 -0.354 0.193 0.587 0.501 ...
  ..$ FC1            : num [1:1920] -0.204 -0.22 0.12 0.365 0.312 ...
  ..$ FC5            : num [1:1920] -1.359 -1.467 0.797 2.431 2.075 ...
  ..$ T7             : num [1:1920] -0.473 -0.51 0.277 0.845 0.722 ...
  ..$ C3             : num [1:1920] -1.022 -1.102 0.599 1.827 1.559 ...
  ..$ CP1            : num [1:1920] -0.0349 -0.0377 0.0205 0.0624 0.0533 ...
  ..$ CP5            : num [1:1920] -1.41 -1.521 0.827 2.521 2.152 ...
  ..$ P7             : num [1:1920] -1.115 -1.203 0.654 1.993 1.701 ...
  ..$ P3             : num [1:1920] -0.912 -0.984 0.535 1.631 1.392 ...
  ..$ Pz             : num [1:1920] 0.0499 0.0538 -0.0292 -0.0891 -0.0761 ...
  ..$ PO3            : num [1:1920] -0.408 -0.44 0.239 0.729 0.623 ...
  ..$ O1             : num [1:1920] -0.954 -1.029 0.559 1.705 1.456 ...
  ..$ Oz             : num [1:1920] 0.247 0.266 -0.145 -0.441 -0.377 ...
  ..$ O2             : num [1:1920] -0.0839 -0.0906 0.0492 0.1501 0.1281 ...
  ..$ PO4            : num [1:1920] -0.0249 -0.0268 0.0146 0.0445 0.038 ...
  ..$ P4             : num [1:1920] -0.431 -0.465 0.253 0.77 0.658 ...
  ..$ P8             : num [1:1920] 0.1376 0.1485 -0.0807 -0.2461 -0.2101 ...
  ..$ CP6            : num [1:1920] -0.664 -0.717 0.39 1.188 1.014 ...
  ..$ CP2            : num [1:1920] 0.448 0.483 -0.263 -0.801 -0.684 ...
  ..$ C4             : num [1:1920] 0.194 0.209 -0.114 -0.347 -0.296 ...
  ..$ FC6            : num [1:1920] 0.913 0.986 -0.536 -1.633 -1.395 ...
  ..$ T8             : num [1:1920] 0.1164 0.1256 -0.0683 -0.2082 -0.1778 ...
  ..$ FC2            : num [1:1920] 1.136 1.226 -0.666 -2.032 -1.734 ...
  ..$ F4             : num [1:1920] 1.11 1.197 -0.651 -1.984 -1.694 ...
  ..$ F8             : num [1:1920] 2.35 2.54 -1.38 -4.21 -3.59 ...
  ..$ AF4            : num [1:1920] 1.95 2.1 -1.14 -3.48 -2.97 ...
  ..$ Fp2            : num [1:1920] 3.68 3.97 -2.16 -6.57 -5.61 ...
  ..$ Fz             : num [1:1920] 1.168 1.26 -0.685 -2.088 -1.782 ...
  ..$ Cz             : num [1:1920] 0.0801 0.0864 -0.047 -0.1431 -0.1222 ...
  ..$ LM             : num [1:1920] 23.6 22.6 21.2 17 13 ...
  ..$ RM             : num [1:1920] -32.2 -33.4 -38.5 -42.1 -44 ...
  ..$ LC             : num [1:1920] 3.82 27.47 48.75 57.74 67.49 ...
  ..$ RC             : num [1:1920] 7.81 23.78 38.38 51.08 64.42 ...
  ..$ TE             : num [1:1920] 51.5 123.9 243.7 374.5 475 ...
  ..$ BE             : num [1:1920] 11.79 1.32 -25.12 -64.49 -91.41 ...
  ..$ UnknownChannel1: num [1:1920] 9.13 10.9 17.36 25.2 31.01 ...
  ..$ UnknownChannel2: num [1:1920] 8.32 10.61 17.42 25.28 31 ...
  ..$ Subject        : chr [1:1920] "21" "21" "21" "21" ...
  ..$ VEOG           : num [1:1920] -39.7 -122.5 -268.8 -439 -566.4 ...
  ..$ HEOG           : num [1:1920] -3.99 3.7 10.38 6.66 3.08 ...
 $ snr.ica  :'data.frame':	32 obs. of  2 variables:
  ..$ Electrode  : Factor w/ 32 levels "AF3","AF4","C3",..: 18 19 1 2 12 10 20 11 13 16 ...
  ..$ MeanSNR.ICA: num [1:32] 13.6 13.5 13.7 11.7 15.5 ...
 $ threshold: num 0.1
 $ S        : num [1:1920, 1:32] 0.481 0.872 0.892 0.83 1.018 ...
  ..- attr(*, "dimnames")=List of 2
  .. ..$ : chr [1:1920] "193" "194" "195" "196" ...
  .. ..$ : NULL
 $ A        : num [1:32, 1:32] 9.75 -6.66 -2.11 -2.28 -1.61 ...
 $ W        : num [1:32, 1:32] -0.1529 0.3483 0.0315 0.0374 -0.0594 ...
 $ K        : num [1:32, 1:32] -0.00447 -0.00493 -0.00262 -0.00276 -0.00157 ...
}
\examples{
data(den)
names(den)
}
\keyword{datasets}

